Rename and fix tests
