+++ AcceptanceTests/Tests/SignUpTests.swift
@@ -1,5 +1,5 @@
 //
-//  SignInTests.swift
+//  SignUpTests.swift
 //  ios-base
 //
 //  Created by Rootstrap on 5/15/17.
@@ -10,18 +10,22 @@ import OHHTTPStubs
 import KIF
 @testable import ios_base
 
-class aSignInTests: KIFTestCase {
+class SignUpTests: KIFTestCase {
   
   override func beforeEach() {
     super.beforeEach()
-
-    if let navigationController = UIApplication.shared.keyWindow?.rootViewController as? UINavigationController {
-      navigationController.popViewController(animated: true)
-    }
     
     tester().waitForView(withAccessibilityIdentifier: "StartView")
-    tester().tapView(withAccessibilityIdentifier: "GoToSignInButton")
-    tester().waitForView(withAccessibilityIdentifier: "SignInView")
+    tester().tapView(withAccessibilityIdentifier: "GoToSignUpButton")
+    tester().waitForView(withAccessibilityIdentifier: "SignUpView")
+  }
+  
+  override func afterEach() {
+    super.afterEach()
+    
+    if let navigationController = UIApplication.shared.keyWindow?.rootViewController as? UINavigationController, !SessionManager.validSession {
+      navigationController.popViewController(animated: true)
+    }
   }
   
   override func afterAll() {
@@ -37,49 +41,56 @@ class aSignInTests: KIFTestCase {
   
   // MARK: - Tests
   
-  func test00SignInEmptyPasswordError() {
-    stub(condition: isPath("/api/v1/users/sign_in")) { _ in
+  func testSignUpEmptyUsernameError() {
+    stub(condition: isPath("/api/v1/users")) { _ in
       return fixture(filePath: "", status: 401, headers: ["Content-Type": "application/json"]).requestTime(0, responseTime: OHHTTPStubsDownloadSpeedWifi)
     }
     
-    tester().enterText("username", intoViewWithAccessibilityIdentifier: "UsernameTextField")
-    tester().tapView(withAccessibilityIdentifier: "SignInButton")
-    showErrorMessage()
-    XCTAssertEqual(SessionManager.validSession, false)
-  }
-  
-  func test01SignInEmptyUsernameError() {
-    stub(condition: isPath("/api/v1/users/sign_in")) { _ in
-      return fixture(filePath: "", status: 401, headers: ["Content-Type": "application/json"]).requestTime(0, responseTime: OHHTTPStubsDownloadSpeedWifi)
-    }
-    
+    tester().enterText("name", intoViewWithAccessibilityIdentifier: "NameTextField")
     tester().enterText("password", intoViewWithAccessibilityIdentifier: "PasswordTextField")
-    tester().tapView(withAccessibilityIdentifier: "SignInButton")
+    tester().enterText("password", intoViewWithAccessibilityIdentifier: "ConfirmPasswordTextField")
+    tester().tapView(withAccessibilityIdentifier: "SignUpButton")
     showErrorMessage()
     XCTAssertEqual(SessionManager.validSession, false)
   }
   
-  func test02SignInEmptyFieldsError() {
-    stub(condition: isPath("/api/v1/users/sign_in")) { _ in
+  func testSignUpMatchPasswordError() {
+    stub(condition: isPath("/api/v1/users")) { _ in
       return fixture(filePath: "", status: 401, headers: ["Content-Type": "application/json"]).requestTime(0, responseTime: OHHTTPStubsDownloadSpeedWifi)
     }
     
-    tester().tapView(withAccessibilityIdentifier: "SignInButton")
+    tester().enterText("name", intoViewWithAccessibilityIdentifier: "NameTextField")
+    tester().enterText("username", intoViewWithAccessibilityIdentifier: "UsernameTextField")
+    tester().enterText("password", intoViewWithAccessibilityIdentifier: "PasswordTextField")
+    tester().enterText("differentPassword", intoViewWithAccessibilityIdentifier: "ConfirmPasswordTextField")
+    tester().tapView(withAccessibilityIdentifier: "SignUpButton")
     showErrorMessage()
     XCTAssertEqual(SessionManager.validSession, false)
   }
   
-  func test03SignInSuccessfully() {
-    stub(condition: isPath("/api/v1/users/sign_in")) { _ in
-      let stubPath = OHPathForFile("SignInSuccessfully.json", type(of: self))
+  func testSignUpEmptyFieldsError() {
+    stub(condition: isPath("/api/v1/users")) { _ in
+      return fixture(filePath: "", status: 401, headers: ["Content-Type": "application/json"]).requestTime(0, responseTime: OHHTTPStubsDownloadSpeedWifi)
+    }
+    
+    tester().tapView(withAccessibilityIdentifier: "SignUpButton")
+    showErrorMessage()
+    XCTAssertEqual(SessionManager.validSession, false)
+  }
+  
+  func testSignUpSuccessfully() {
+    stub(condition: isPath("/api/v1/users")) { _ in
+      let stubPath = OHPathForFile("SignUpSuccessfully.json", type(of: self))
       return fixture(filePath: stubPath!, status: 200, headers: ["Content-Type": "application/json",
                                                                  "uid": "rootstrap@gmail.com",
                                                                  "client": "client", "access-token": "token"]).requestTime(0, responseTime: OHHTTPStubsDownloadSpeedWifi)
     }
     
-    tester().enterText("rootstrap@gmail.com", intoViewWithAccessibilityIdentifier: "UsernameTextField")
-    tester().enterText("123456789", intoViewWithAccessibilityIdentifier: "PasswordTextField")
-    tester().tapView(withAccessibilityIdentifier: "SignInButton")
+    tester().enterText("name", intoViewWithAccessibilityIdentifier: "NameTextField")
+    tester().enterText("username", intoViewWithAccessibilityIdentifier: "UsernameTextField")
+    tester().enterText("password", intoViewWithAccessibilityIdentifier: "PasswordTextField")
+    tester().enterText("password", intoViewWithAccessibilityIdentifier: "ConfirmPasswordTextField")
+    tester().tapView(withAccessibilityIdentifier: "SignUpButton")
     tester().waitForView(withAccessibilityIdentifier: "AfterLoginSignupView")
     XCTAssertEqual(SessionManager.validSession, true)
   }
